refactor: extract canvas and panels
